+++ bindings/go/scip/sort.go
@@ -33,14 +33,14 @@ func FindOccurrences(occurrences []*Occurrence, targetLine, targetCharacter int3
 	var filtered []*Occurrence
 	pos := Position{targetLine, targetCharacter}
 	for _, occurrence := range occurrences {
-		if NewRange(occurrence.Range).Contains(pos) {
+		if NewRangeUnchecked(occurrence.Range).Contains(pos) {
 			filtered = append(filtered, occurrence)
 		}
 	}
 
 	sort.Slice(filtered, func(i, j int) bool {
 		// Ordered so that the least precise (largest) range comes last
-		return NewRange(filtered[i].Range).CompareStrict(NewRange(filtered[j].Range)) > 0
+		return NewRangeUnchecked(filtered[i].Range).CompareStrict(NewRangeUnchecked(filtered[j].Range)) > 0
 	})
 
 	return filtered
@@ -52,8 +52,8 @@ func FindOccurrences(occurrences []*Occurrence, targetLine, targetCharacter int3
 // occurrences are sorted by symbol name.
 func SortOccurrences(occurrences []*Occurrence) []*Occurrence {
 	sort.Slice(occurrences, func(i, j int) bool {
-		r1 := NewRange(occurrences[i].Range)
-		r2 := NewRange(occurrences[j].Range)
+		r1 := NewRangeUnchecked(occurrences[i].Range)
+		r2 := NewRangeUnchecked(occurrences[j].Range)
 		if ret := r1.CompareStrict(r2); ret != 0 {
 			return ret < 0
 		}
@@ -75,8 +75,8 @@ func rawRangesEqual(a, b []int32) bool {
 		return true
 	}
 
-	ra := NewRange(a)
-	rb := NewRange(b)
+	ra := NewRangeUnchecked(a)
+	rb := NewRangeUnchecked(b)
 
 	return ra.Start.Line == rb.Start.Line && ra.Start.Character == rb.Start.Character && ra.End.Line == rb.End.Line && ra.End.Character == rb.End.Character
 }